Reorganized repository root.
