+++ package.json
@@ -3,14 +3,14 @@
   "version": "1.3.0",
   "description": "Kinto Web Administration Console in React.js",
   "scripts": {
-    "build:gh-pages": "rimraf gh-pages && NODE_ENV=production node_modules/.bin/webpack --progress --config webpack.prod.js && cp index.prod.html gh-pages/index.html",
+    "build:gh-pages": "rimraf gh-pages && NODE_ENV=production node_modules/.bin/webpack --progress --config webpack.prod.js && cp html/prod.html gh-pages/index.html",
     "build:lib": "rimraf lib && NODE_ENV=production babel -d lib/ src/",
     "build:readme": "toctoc README.md -w",
     "flow-check": "node_modules/.bin/flow check",
     "lint": "node_modules/.bin/eslint src test test-browser",
     "publish-to-gh-pages": "npm run build:gh-pages && gh-pages --add --dist gh-pages/",
     "publish-to-npm": "npm run build:lib && npm publish",
-    "start": "NODE_ENV=development node devServer.js",
+    "start": "NODE_ENV=development node server/dev.js",
     "tdd": "NODE_ENV=test node_modules/.bin/babel-node node_modules/.bin/_mocha --require ignore-styles --require ./test/setup-jsdom.js 'test/**/*_test.js' --watch",
     "tdd-browser": "NODE_ENV=test node_modules/.bin/babel-node node_modules/.bin/_mocha 'test-browser/**/*_test.js' --watch",
     "test": "NODE_ENV=test node_modules/.bin/babel-node node_modules/.bin/_mocha --require ignore-styles --require ./test/setup-jsdom.js 'test/**/*_test.js'",
