+++ server/dev.js
@@ -1,7 +1,7 @@
 var path = require("path");
 var express = require("express");
 var webpack = require("webpack");
-var config = require("./webpack.dev");
+var config = require("../webpack.dev");
 
 var app = express();
 var compiler = webpack(config);
@@ -15,7 +15,7 @@ app.use(require("webpack-dev-middleware")(compiler, {
 app.use(require("webpack-hot-middleware")(compiler));
 
 app.get("*", function(req, res) {
-  res.sendFile(path.join(__dirname, "index.dev.html"));
+  res.sendFile(path.join(__dirname, "..", "html", "dev.html"));
 });
 
 app.listen(3000, "0.0.0.0", function(err) {
