+++ server/test.js
@@ -2,7 +2,7 @@ var path = require("path");
 var express = require("express");
 
 var app = express();
-app.use(express.static(path.join(__dirname, "gh-pages")));
+app.use(express.static(path.join(__dirname, "..", "gh-pages")));
 
 let server;
